LayoutTests/platform/gtk/imported/w3c/web-platform-tests/cookies/attributes/attributes-ctl.sub-expected.txt


PASS Cookie with %x0 in Domain attribute value is handled correctly.
PASS Cookie with %x0 after Domain attribute value is handled correctly.
PASS Cookie with %x0 in Path attribute value is handled correctly.
FAIL Cookie with %x0 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0path2=t"
FAIL Cookie with %x0 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0maxage=t"
FAIL Cookie with %x0 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0maxage2=t"
FAIL Cookie with %x0 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0expires=t"
FAIL Cookie with %x0 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0expires2=t"
FAIL Cookie with %x0 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0secure=t"
PASS Cookie with %x0 after Secure attribute is handled correctly.
FAIL Cookie with %x0 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0httponly=t"
FAIL Cookie with %x0 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0samesite=t"
FAIL Cookie with %x0 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test0samesite2=t"
PASS Cookie with %x1 in Domain attribute value is handled correctly.
PASS Cookie with %x1 after Domain attribute value is handled correctly.
PASS Cookie with %x1 in Path attribute value is handled correctly.
FAIL Cookie with %x1 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1path2=t"
FAIL Cookie with %x1 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1maxage=t"
FAIL Cookie with %x1 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1maxage2=t"
FAIL Cookie with %x1 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1expires=t"
FAIL Cookie with %x1 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1expires2=t"
FAIL Cookie with %x1 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1secure=t"
PASS Cookie with %x1 after Secure attribute is handled correctly.
FAIL Cookie with %x1 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1httponly=t"
FAIL Cookie with %x1 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1samesite=t"
FAIL Cookie with %x1 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test1samesite2=t"
PASS Cookie with %x2 in Domain attribute value is handled correctly.
PASS Cookie with %x2 after Domain attribute value is handled correctly.
PASS Cookie with %x2 in Path attribute value is handled correctly.
FAIL Cookie with %x2 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2path2=t"
FAIL Cookie with %x2 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2maxage=t"
FAIL Cookie with %x2 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2maxage2=t"
FAIL Cookie with %x2 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2expires=t"
FAIL Cookie with %x2 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2expires2=t"
FAIL Cookie with %x2 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2secure=t"
PASS Cookie with %x2 after Secure attribute is handled correctly.
FAIL Cookie with %x2 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2httponly=t"
FAIL Cookie with %x2 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2samesite=t"
FAIL Cookie with %x2 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test2samesite2=t"
PASS Cookie with %x3 in Domain attribute value is handled correctly.
PASS Cookie with %x3 after Domain attribute value is handled correctly.
PASS Cookie with %x3 in Path attribute value is handled correctly.
FAIL Cookie with %x3 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3path2=t"
FAIL Cookie with %x3 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3maxage=t"
FAIL Cookie with %x3 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3maxage2=t"
FAIL Cookie with %x3 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3expires=t"
FAIL Cookie with %x3 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3expires2=t"
FAIL Cookie with %x3 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3secure=t"
PASS Cookie with %x3 after Secure attribute is handled correctly.
FAIL Cookie with %x3 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3httponly=t"
FAIL Cookie with %x3 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3samesite=t"
FAIL Cookie with %x3 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test3samesite2=t"
PASS Cookie with %x4 in Domain attribute value is handled correctly.
PASS Cookie with %x4 after Domain attribute value is handled correctly.
PASS Cookie with %x4 in Path attribute value is handled correctly.
FAIL Cookie with %x4 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4path2=t"
FAIL Cookie with %x4 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4maxage=t"
FAIL Cookie with %x4 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4maxage2=t"
FAIL Cookie with %x4 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4expires=t"
FAIL Cookie with %x4 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4expires2=t"
FAIL Cookie with %x4 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4secure=t"
PASS Cookie with %x4 after Secure attribute is handled correctly.
FAIL Cookie with %x4 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4httponly=t"
FAIL Cookie with %x4 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4samesite=t"
FAIL Cookie with %x4 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test4samesite2=t"
PASS Cookie with %x5 in Domain attribute value is handled correctly.
PASS Cookie with %x5 after Domain attribute value is handled correctly.
PASS Cookie with %x5 in Path attribute value is handled correctly.
FAIL Cookie with %x5 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5path2=t"
FAIL Cookie with %x5 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5maxage=t"
FAIL Cookie with %x5 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5maxage2=t"
FAIL Cookie with %x5 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5expires=t"
FAIL Cookie with %x5 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5expires2=t"
FAIL Cookie with %x5 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5secure=t"
PASS Cookie with %x5 after Secure attribute is handled correctly.
FAIL Cookie with %x5 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5httponly=t"
FAIL Cookie with %x5 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5samesite=t"
FAIL Cookie with %x5 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test5samesite2=t"
PASS Cookie with %x6 in Domain attribute value is handled correctly.
PASS Cookie with %x6 after Domain attribute value is handled correctly.
PASS Cookie with %x6 in Path attribute value is handled correctly.
FAIL Cookie with %x6 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6path2=t"
FAIL Cookie with %x6 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6maxage=t"
FAIL Cookie with %x6 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6maxage2=t"
FAIL Cookie with %x6 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6expires=t"
FAIL Cookie with %x6 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6expires2=t"
FAIL Cookie with %x6 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6secure=t"
PASS Cookie with %x6 after Secure attribute is handled correctly.
FAIL Cookie with %x6 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6httponly=t"
FAIL Cookie with %x6 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6samesite=t"
FAIL Cookie with %x6 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test6samesite2=t"
PASS Cookie with %x7 in Domain attribute value is handled correctly.
PASS Cookie with %x7 after Domain attribute value is handled correctly.
PASS Cookie with %x7 in Path attribute value is handled correctly.
FAIL Cookie with %x7 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7path2=t"
FAIL Cookie with %x7 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7maxage=t"
FAIL Cookie with %x7 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7maxage2=t"
FAIL Cookie with %x7 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7expires=t"
FAIL Cookie with %x7 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7expires2=t"
FAIL Cookie with %x7 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7secure=t"
PASS Cookie with %x7 after Secure attribute is handled correctly.
FAIL Cookie with %x7 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7httponly=t"
FAIL Cookie with %x7 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7samesite=t"
FAIL Cookie with %x7 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test7samesite2=t"
PASS Cookie with %x8 in Domain attribute value is handled correctly.
PASS Cookie with %x8 after Domain attribute value is handled correctly.
PASS Cookie with %x8 in Path attribute value is handled correctly.
FAIL Cookie with %x8 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8path2=t"
FAIL Cookie with %x8 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8maxage=t"
FAIL Cookie with %x8 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8maxage2=t"
FAIL Cookie with %x8 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8expires=t"
FAIL Cookie with %x8 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8expires2=t"
FAIL Cookie with %x8 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8secure=t"
PASS Cookie with %x8 after Secure attribute is handled correctly.
FAIL Cookie with %x8 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8httponly=t"
FAIL Cookie with %x8 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8samesite=t"
FAIL Cookie with %x8 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test8samesite2=t"
FAIL Cookie with %x9 in Domain attribute value is handled correctly. assert_equals: The cookie was set as expected. expected "test9domain=t" but got ""
FAIL Cookie with %x9 after Domain attribute value is handled correctly. assert_equals: The cookie was set as expected. expected "test9domain2=t" but got ""
FAIL Cookie with %x9 in Path attribute value is handled correctly. assert_equals: The cookie was set as expected. expected "test9path=t" but got ""
PASS Cookie with %x9 after Path attribute value is handled correctly.
PASS Cookie with %x9 in Max-Age attribute value is handled correctly.
PASS Cookie with %x9 after Max-Age attribute value is handled correctly.
PASS Cookie with %x9 in Expires attribute value is handled correctly.
PASS Cookie with %x9 after Expires attribute value is handled correctly.
PASS Cookie with %x9 in Secure attribute is handled correctly.
FAIL Cookie with %x9 after Secure attribute is handled correctly. assert_equals: The cookie was set as expected. expected "test9secure2=t" but got ""
PASS Cookie with %x9 in HttpOnly attribute is handled correctly.
PASS Cookie with %x9 in SameSite attribute value is handled correctly.
PASS Cookie with %x9 after SameSite attribute value is handled correctly.
PASS Cookie with %xa in Domain attribute value is handled correctly.
PASS Cookie with %xa after Domain attribute value is handled correctly.
FAIL Cookie with %xa in Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "st"
FAIL Cookie with %xa after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10path2=t"
FAIL Cookie with %xa in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10maxage=t; 00"
FAIL Cookie with %xa after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10maxage2=t"
FAIL Cookie with %xa in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10expires=t; 38 00:00:00 GMT"
FAIL Cookie with %xa after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10expires2=t"
FAIL Cookie with %xa in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10secure=t; ure"
PASS Cookie with %xa after Secure attribute is handled correctly.
FAIL Cookie with %xa in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10httponly=t; Only"
FAIL Cookie with %xa in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10samesite=t; x"
FAIL Cookie with %xa after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test10samesite2=t"
PASS Cookie with %xb in Domain attribute value is handled correctly.
PASS Cookie with %xb after Domain attribute value is handled correctly.
PASS Cookie with %xb in Path attribute value is handled correctly.
FAIL Cookie with %xb after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11path2=t"
FAIL Cookie with %xb in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11maxage=t"
FAIL Cookie with %xb after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11maxage2=t"
FAIL Cookie with %xb in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11expires=t"
FAIL Cookie with %xb after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11expires2=t"
FAIL Cookie with %xb in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11secure=t"
PASS Cookie with %xb after Secure attribute is handled correctly.
FAIL Cookie with %xb in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11httponly=t"
FAIL Cookie with %xb in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11samesite=t"
FAIL Cookie with %xb after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test11samesite2=t"
PASS Cookie with %xc in Domain attribute value is handled correctly.
PASS Cookie with %xc after Domain attribute value is handled correctly.
PASS Cookie with %xc in Path attribute value is handled correctly.
FAIL Cookie with %xc after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12path2=t"
FAIL Cookie with %xc in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12maxage=t"
FAIL Cookie with %xc after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12maxage2=t"
FAIL Cookie with %xc in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12expires=t"
FAIL Cookie with %xc after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12expires2=t"
FAIL Cookie with %xc in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12secure=t"
PASS Cookie with %xc after Secure attribute is handled correctly.
FAIL Cookie with %xc in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12httponly=t"
FAIL Cookie with %xc in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12samesite=t"
FAIL Cookie with %xc after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test12samesite2=t"
PASS Cookie with %xd in Domain attribute value is handled correctly.
PASS Cookie with %xd after Domain attribute value is handled correctly.
PASS Cookie with %xd in Path attribute value is handled correctly.
FAIL Cookie with %xd after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13path2=t"
FAIL Cookie with %xd in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13maxage=t"
FAIL Cookie with %xd after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13maxage2=t"
FAIL Cookie with %xd in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13expires=t"
FAIL Cookie with %xd after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13expires2=t"
FAIL Cookie with %xd in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13secure=t"
PASS Cookie with %xd after Secure attribute is handled correctly.
FAIL Cookie with %xd in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13httponly=t"
FAIL Cookie with %xd in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13samesite=t"
FAIL Cookie with %xd after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test13samesite2=t"
PASS Cookie with %xe in Domain attribute value is handled correctly.
PASS Cookie with %xe after Domain attribute value is handled correctly.
PASS Cookie with %xe in Path attribute value is handled correctly.
FAIL Cookie with %xe after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14path2=t"
FAIL Cookie with %xe in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14maxage=t"
FAIL Cookie with %xe after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14maxage2=t"
FAIL Cookie with %xe in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14expires=t"
FAIL Cookie with %xe after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14expires2=t"
FAIL Cookie with %xe in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14secure=t"
PASS Cookie with %xe after Secure attribute is handled correctly.
FAIL Cookie with %xe in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14httponly=t"
FAIL Cookie with %xe in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14samesite=t"
FAIL Cookie with %xe after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test14samesite2=t"
PASS Cookie with %xf in Domain attribute value is handled correctly.
PASS Cookie with %xf after Domain attribute value is handled correctly.
PASS Cookie with %xf in Path attribute value is handled correctly.
FAIL Cookie with %xf after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15path2=t"
FAIL Cookie with %xf in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15maxage=t"
FAIL Cookie with %xf after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15maxage2=t"
FAIL Cookie with %xf in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15expires=t"
FAIL Cookie with %xf after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15expires2=t"
FAIL Cookie with %xf in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15secure=t"
PASS Cookie with %xf after Secure attribute is handled correctly.
FAIL Cookie with %xf in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15httponly=t"
FAIL Cookie with %xf in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15samesite=t"
FAIL Cookie with %xf after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test15samesite2=t"
PASS Cookie with %x10 in Domain attribute value is handled correctly.
PASS Cookie with %x10 after Domain attribute value is handled correctly.
PASS Cookie with %x10 in Path attribute value is handled correctly.
FAIL Cookie with %x10 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16path2=t"
FAIL Cookie with %x10 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16maxage=t"
FAIL Cookie with %x10 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16maxage2=t"
FAIL Cookie with %x10 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16expires=t"
FAIL Cookie with %x10 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16expires2=t"
FAIL Cookie with %x10 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16secure=t"
PASS Cookie with %x10 after Secure attribute is handled correctly.
FAIL Cookie with %x10 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16httponly=t"
FAIL Cookie with %x10 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16samesite=t"
FAIL Cookie with %x10 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test16samesite2=t"
PASS Cookie with %x11 in Domain attribute value is handled correctly.
PASS Cookie with %x11 after Domain attribute value is handled correctly.
PASS Cookie with %x11 in Path attribute value is handled correctly.
FAIL Cookie with %x11 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17path2=t"
FAIL Cookie with %x11 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17maxage=t"
FAIL Cookie with %x11 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17maxage2=t"
FAIL Cookie with %x11 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17expires=t"
FAIL Cookie with %x11 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17expires2=t"
FAIL Cookie with %x11 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17secure=t"
PASS Cookie with %x11 after Secure attribute is handled correctly.
FAIL Cookie with %x11 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17httponly=t"
FAIL Cookie with %x11 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17samesite=t"
FAIL Cookie with %x11 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test17samesite2=t"
PASS Cookie with %x12 in Domain attribute value is handled correctly.
PASS Cookie with %x12 after Domain attribute value is handled correctly.
PASS Cookie with %x12 in Path attribute value is handled correctly.
FAIL Cookie with %x12 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18path2=t"
FAIL Cookie with %x12 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18maxage=t"
FAIL Cookie with %x12 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18maxage2=t"
FAIL Cookie with %x12 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18expires=t"
FAIL Cookie with %x12 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18expires2=t"
FAIL Cookie with %x12 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18secure=t"
PASS Cookie with %x12 after Secure attribute is handled correctly.
FAIL Cookie with %x12 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18httponly=t"
FAIL Cookie with %x12 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18samesite=t"
FAIL Cookie with %x12 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test18samesite2=t"
PASS Cookie with %x13 in Domain attribute value is handled correctly.
PASS Cookie with %x13 after Domain attribute value is handled correctly.
PASS Cookie with %x13 in Path attribute value is handled correctly.
FAIL Cookie with %x13 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19path2=t"
FAIL Cookie with %x13 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19maxage=t"
FAIL Cookie with %x13 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19maxage2=t"
FAIL Cookie with %x13 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19expires=t"
FAIL Cookie with %x13 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19expires2=t"
FAIL Cookie with %x13 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19secure=t"
PASS Cookie with %x13 after Secure attribute is handled correctly.
FAIL Cookie with %x13 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19httponly=t"
FAIL Cookie with %x13 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19samesite=t"
FAIL Cookie with %x13 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test19samesite2=t"
PASS Cookie with %x14 in Domain attribute value is handled correctly.
PASS Cookie with %x14 after Domain attribute value is handled correctly.
PASS Cookie with %x14 in Path attribute value is handled correctly.
FAIL Cookie with %x14 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20path2=t"
FAIL Cookie with %x14 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20maxage=t"
FAIL Cookie with %x14 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20maxage2=t"
FAIL Cookie with %x14 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20expires=t"
FAIL Cookie with %x14 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20expires2=t"
FAIL Cookie with %x14 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20secure=t"
PASS Cookie with %x14 after Secure attribute is handled correctly.
FAIL Cookie with %x14 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20httponly=t"
FAIL Cookie with %x14 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20samesite=t"
FAIL Cookie with %x14 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test20samesite2=t"
PASS Cookie with %x15 in Domain attribute value is handled correctly.
PASS Cookie with %x15 after Domain attribute value is handled correctly.
PASS Cookie with %x15 in Path attribute value is handled correctly.
FAIL Cookie with %x15 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21path2=t"
FAIL Cookie with %x15 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21maxage=t"
FAIL Cookie with %x15 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21maxage2=t"
FAIL Cookie with %x15 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21expires=t"
FAIL Cookie with %x15 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21expires2=t"
FAIL Cookie with %x15 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21secure=t"
PASS Cookie with %x15 after Secure attribute is handled correctly.
FAIL Cookie with %x15 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21httponly=t"
FAIL Cookie with %x15 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21samesite=t"
FAIL Cookie with %x15 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test21samesite2=t"
PASS Cookie with %x16 in Domain attribute value is handled correctly.
PASS Cookie with %x16 after Domain attribute value is handled correctly.
PASS Cookie with %x16 in Path attribute value is handled correctly.
FAIL Cookie with %x16 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22path2=t"
FAIL Cookie with %x16 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22maxage=t"
FAIL Cookie with %x16 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22maxage2=t"
FAIL Cookie with %x16 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22expires=t"
FAIL Cookie with %x16 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22expires2=t"
FAIL Cookie with %x16 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22secure=t"
PASS Cookie with %x16 after Secure attribute is handled correctly.
FAIL Cookie with %x16 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22httponly=t"
FAIL Cookie with %x16 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22samesite=t"
FAIL Cookie with %x16 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test22samesite2=t"
PASS Cookie with %x17 in Domain attribute value is handled correctly.
PASS Cookie with %x17 after Domain attribute value is handled correctly.
PASS Cookie with %x17 in Path attribute value is handled correctly.
FAIL Cookie with %x17 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23path2=t"
FAIL Cookie with %x17 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23maxage=t"
FAIL Cookie with %x17 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23maxage2=t"
FAIL Cookie with %x17 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23expires=t"
FAIL Cookie with %x17 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23expires2=t"
FAIL Cookie with %x17 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23secure=t"
PASS Cookie with %x17 after Secure attribute is handled correctly.
FAIL Cookie with %x17 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23httponly=t"
FAIL Cookie with %x17 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23samesite=t"
FAIL Cookie with %x17 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test23samesite2=t"
PASS Cookie with %x18 in Domain attribute value is handled correctly.
PASS Cookie with %x18 after Domain attribute value is handled correctly.
PASS Cookie with %x18 in Path attribute value is handled correctly.
FAIL Cookie with %x18 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24path2=t"
FAIL Cookie with %x18 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24maxage=t"
FAIL Cookie with %x18 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24maxage2=t"
FAIL Cookie with %x18 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24expires=t"
FAIL Cookie with %x18 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24expires2=t"
FAIL Cookie with %x18 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24secure=t"
PASS Cookie with %x18 after Secure attribute is handled correctly.
FAIL Cookie with %x18 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24httponly=t"
FAIL Cookie with %x18 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24samesite=t"
FAIL Cookie with %x18 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test24samesite2=t"
PASS Cookie with %x19 in Domain attribute value is handled correctly.
PASS Cookie with %x19 after Domain attribute value is handled correctly.
PASS Cookie with %x19 in Path attribute value is handled correctly.
FAIL Cookie with %x19 after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25path2=t"
FAIL Cookie with %x19 in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25maxage=t"
FAIL Cookie with %x19 after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25maxage2=t"
FAIL Cookie with %x19 in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25expires=t"
FAIL Cookie with %x19 after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25expires2=t"
FAIL Cookie with %x19 in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25secure=t"
PASS Cookie with %x19 after Secure attribute is handled correctly.
FAIL Cookie with %x19 in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25httponly=t"
FAIL Cookie with %x19 in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25samesite=t"
FAIL Cookie with %x19 after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test25samesite2=t"
PASS Cookie with %x1a in Domain attribute value is handled correctly.
PASS Cookie with %x1a after Domain attribute value is handled correctly.
PASS Cookie with %x1a in Path attribute value is handled correctly.
FAIL Cookie with %x1a after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26path2=t"
FAIL Cookie with %x1a in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26maxage=t"
FAIL Cookie with %x1a after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26maxage2=t"
FAIL Cookie with %x1a in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26expires=t"
FAIL Cookie with %x1a after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26expires2=t"
FAIL Cookie with %x1a in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26secure=t"
PASS Cookie with %x1a after Secure attribute is handled correctly.
FAIL Cookie with %x1a in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26httponly=t"
FAIL Cookie with %x1a in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26samesite=t"
FAIL Cookie with %x1a after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test26samesite2=t"
PASS Cookie with %x1b in Domain attribute value is handled correctly.
PASS Cookie with %x1b after Domain attribute value is handled correctly.
PASS Cookie with %x1b in Path attribute value is handled correctly.
FAIL Cookie with %x1b after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27path2=t"
FAIL Cookie with %x1b in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27maxage=t"
FAIL Cookie with %x1b after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27maxage2=t"
FAIL Cookie with %x1b in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27expires=t"
FAIL Cookie with %x1b after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27expires2=t"
FAIL Cookie with %x1b in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27secure=t"
PASS Cookie with %x1b after Secure attribute is handled correctly.
FAIL Cookie with %x1b in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27httponly=t"
FAIL Cookie with %x1b in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27samesite=t"
FAIL Cookie with %x1b after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test27samesite2=t"
PASS Cookie with %x1c in Domain attribute value is handled correctly.
PASS Cookie with %x1c after Domain attribute value is handled correctly.
PASS Cookie with %x1c in Path attribute value is handled correctly.
FAIL Cookie with %x1c after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28path2=t"
FAIL Cookie with %x1c in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28maxage=t"
FAIL Cookie with %x1c after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28maxage2=t"
FAIL Cookie with %x1c in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28expires=t"
FAIL Cookie with %x1c after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28expires2=t"
FAIL Cookie with %x1c in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28secure=t"
PASS Cookie with %x1c after Secure attribute is handled correctly.
FAIL Cookie with %x1c in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28httponly=t"
FAIL Cookie with %x1c in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28samesite=t"
FAIL Cookie with %x1c after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test28samesite2=t"
PASS Cookie with %x1d in Domain attribute value is handled correctly.
PASS Cookie with %x1d after Domain attribute value is handled correctly.
PASS Cookie with %x1d in Path attribute value is handled correctly.
FAIL Cookie with %x1d after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29path2=t"
FAIL Cookie with %x1d in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29maxage=t"
FAIL Cookie with %x1d after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29maxage2=t"
FAIL Cookie with %x1d in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29expires=t"
FAIL Cookie with %x1d after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29expires2=t"
FAIL Cookie with %x1d in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29secure=t"
PASS Cookie with %x1d after Secure attribute is handled correctly.
FAIL Cookie with %x1d in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29httponly=t"
FAIL Cookie with %x1d in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29samesite=t"
FAIL Cookie with %x1d after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test29samesite2=t"
PASS Cookie with %x1e in Domain attribute value is handled correctly.
PASS Cookie with %x1e after Domain attribute value is handled correctly.
PASS Cookie with %x1e in Path attribute value is handled correctly.
FAIL Cookie with %x1e after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30path2=t"
FAIL Cookie with %x1e in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30maxage=t"
FAIL Cookie with %x1e after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30maxage2=t"
FAIL Cookie with %x1e in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30expires=t"
FAIL Cookie with %x1e after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30expires2=t"
FAIL Cookie with %x1e in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30secure=t"
PASS Cookie with %x1e after Secure attribute is handled correctly.
FAIL Cookie with %x1e in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30httponly=t"
FAIL Cookie with %x1e in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30samesite=t"
FAIL Cookie with %x1e after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test30samesite2=t"
PASS Cookie with %x1f in Domain attribute value is handled correctly.
PASS Cookie with %x1f after Domain attribute value is handled correctly.
PASS Cookie with %x1f in Path attribute value is handled correctly.
FAIL Cookie with %x1f after Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31path2=t"
FAIL Cookie with %x1f in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31maxage=t"
FAIL Cookie with %x1f after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31maxage2=t"
FAIL Cookie with %x1f in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31expires=t"
FAIL Cookie with %x1f after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31expires2=t"
FAIL Cookie with %x1f in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31secure=t"
PASS Cookie with %x1f after Secure attribute is handled correctly.
FAIL Cookie with %x1f in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31httponly=t"
FAIL Cookie with %x1f in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31samesite=t"
FAIL Cookie with %x1f after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test31samesite2=t"
PASS Cookie with %x7f in Domain attribute value is handled correctly.
PASS Cookie with %x7f after Domain attribute value is handled correctly.
FAIL Cookie with %x7f in Path attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127path=t"
PASS Cookie with %x7f after Path attribute value is handled correctly.
FAIL Cookie with %x7f in Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127maxage=t"
FAIL Cookie with %x7f after Max-Age attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127maxage2=t"
FAIL Cookie with %x7f in Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127expires=t"
FAIL Cookie with %x7f after Expires attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127expires2=t"
FAIL Cookie with %x7f in Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127secure=t"
FAIL Cookie with %x7f after Secure attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127secure2=t"
FAIL Cookie with %x7f in HttpOnly attribute is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127httponly=t"
FAIL Cookie with %x7f in SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127samesite=t"
FAIL Cookie with %x7f after SameSite attribute value is handled correctly. assert_equals: The cookie was rejected. expected "" but got "test127samesite2=t"
